+++ src/main/java/universite_paris8/iut/rdias/towerdefense/model/SorcererTower.java
@@ -4,7 +4,7 @@ public class SorcererTower extends Tower {
     private int dmgPerSec;
     private int zoneDuration;
 
-    public SorcererTower(Environnement env, int tHp, int tDmg, int tId, double tRange, double tX, double tY, double tSpeedAttack, int sTdmgPerSec, int sCost, int sTZoneDuration, int sDmgPersec) {
+    public SorcererTower(Environnement env, int tHp, int tDmg, int tId, double tRange, double tX, double tY, int tSpeedAttack, int sTdmgPerSec, int sCost, int sTZoneDuration, int sDmgPersec) {
         super(env, tHp, tDmg, tId, tRange, tX, tY, tSpeedAttack, sCost);
         this.dmgPerSec = sTdmgPerSec;
         this.zoneDuration = sTZoneDuration;
